Lexishift-main/requirements.txt
Flask==2.0.1
reportlab==3.5.67
pytesseract==0.3.8
Pillow==8.2.0
PyPDF2==1.26.0
scikit-learn==0.24.2
pandas==1.2.4
numpy==1.20.3
